Add oppo feature builder
